+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - fix permission for classroom creation
 - fix permission for filedepository creation
+- fix permission for markdown document creation
 
 ## [4.0.0-beta.10] - 2022-10-19